# componente-a/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: API Componente A - Clientes y Pedidos
  version: 1.0.0
  description: API para gestión de clientes y pedidos con MariaDB
  contact:
    name: Soporte MultiPedidos
    email: soporte@multipedidos.com

servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo

paths:
  /clientes:
    post:
      operationId: crearCliente
      summary: Crear un cliente
      tags:
        - Clientes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteInput'
      responses:
        '201':
          description: Cliente creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '400':
          description: Datos inválidos
        '500':
          description: Error interno del servidor
          
    get:
      operationId: listarClientes
      summary: Listar todos los clientes
      tags:
        - Clientes
      responses:
        '200':
          description: Lista de clientes obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'
        '500':
          description: Error interno del servidor

  /clientes/{id}:
    get:
      operationId: obtenerCliente
      summary: Obtener un cliente por ID
      tags:
        - Clientes
      parameters:
        - name: id
          in: path
          required: true
          description: ID del cliente
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '404':
          description: Cliente no encontrado
        '500':
          description: Error interno del servidor

  /pedidos:
    post:
      operationId: crearPedido
      summary: Crear un pedido
      tags:
        - Pedidos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoInput'
      responses:
        '201':
          description: Pedido creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '400':
          description: Datos inválidos
        '500':
          description: Error interno del servidor
          
    get:
      operationId: listarPedidos
      summary: Listar todos los pedidos
      tags:
        - Pedidos
      responses:
        '200':
          description: Lista de pedidos obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pedido'
        '500':
          description: Error interno del servidor

  /pedidos/{id}:
    get:
      operationId: obtenerPedido
      summary: Obtener un pedido por ID
      tags:
        - Pedidos
      parameters:
        - name: id
          in: path
          required: true
          description: ID del pedido
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '404':
          description: Pedido no encontrado
        '500':
          description: Error interno del servidor

components:
  schemas:
    Cliente:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: "Juan Pérez"
        correo:
          type: string
          format: email
          example: "juan@example.com"
        fechaCreacion:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
      required:
        - id
        - nombre
        - correo
        
    ClienteInput:
      type: object
      properties:
        nombre:
          type: string
          example: "Juan Pérez"
        correo:
          type: string
          format: email
          example: "juan@example.com"
      required:
        - nombre
        - correo
        
    Producto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: "Laptop Gaming"
        precio:
          type: number
          format: double
          example: 1200.50
        descripcion:
          type: string
          example: "Laptop para gaming de alta gama"
      required:
        - id
        - nombre
        - precio
        
    ItemPedido:
      type: object
      properties:
        productoId:
          type: integer
          format: int64
          example: 1
        cantidad:
          type: integer
          example: 2
        precioUnitario:
          type: number
          format: double
          example: 1200.50
        subtotal:
          type: number
          format: double
          example: 2401.00
      required:
        - productoId
        - cantidad
        - precioUnitario
        - subtotal
        
    Pedido:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        clienteId:
          type: integer
          format: int64
          example: 1
        clienteNombre:
          type: string
          example: "Juan Pérez"
        productos:
          type: array
          items:
            $ref: '#/components/schemas/ItemPedido'
        subtotal:
          type: number
          format: double
          example: 2401.00
        descuento:
          type: number
          format: double
          example: 192.08
        total:
          type: number
          format: double
          example: 2473.83
        estado:
          type: string
          example: "PENDIENTE"
        fechaCreacion:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
      required:
        - id
        - clienteId
        - productos
        - subtotal
        - descuento
        - total
        - estado
        
    PedidoInput:
      type: object
      properties:
        clienteId:
          type: integer
          format: int64
          example: 1
        productos:
          type: array
          items:
            $ref: '#/components/schemas/ItemPedidoInput'
      required:
        - clienteId
        - productos
        
    ItemPedidoInput:
      type: object
      properties:
        productoId:
          type: integer
          format: int64
          example: 1
        cantidad:
          type: integer
          example: 2
      required:
        - productoId
        - cantidad

  responses:
    NotFound:
      description: Recurso no encontrado
    BadRequest:
      description: Solicitud inválida
    ServerError:
      description: Error interno del servidor